部分

+++ 97【考试题型：数组】重复元素/noj097.c
@@ -5,7 +5,7 @@
 
 
 struct number{
-    int data;
+    long long data;
     int sign;
 };
 
@@ -19,7 +19,7 @@ int main(){
     scanf("%d",&n);
     int a[n];
     for(int i=0;i<n;i++){
-        scanf("%d",&a[i]);
+        scanf("%lld",&a[i]);
     }
     int k=0;
